refactor: extract JobHeader; reuse on task list page
JobHeader.svelte holds the dark job-detail banner (title, customer,
status row, P&L grid). JobDetail uses it; JobTaskListPage now uses it
too, replacing the previous lightweight h2 / status-line block. The
back link moves below the header and is renamed "back to overview".
